refactoring login and registration + updating tests
