+++ src/components/login/login.test.js
@@ -106,42 +106,6 @@ describe('Login', () => {
     });
   });
 
-  describe('validateUrl', () => {
-    beforeEach('', () => {
-      wrapper = shallow(<Login {...props}/>, options);
-    });
-
-    it('should set address and addressValidity="" for a valid address', () => {
-      const validURL = 'http://localhost:8080';
-      const data = wrapper.instance().validateUrl(validURL);
-      const expectedData = {
-        address: validURL,
-        addressValidity: '',
-      };
-      expect(data).to.deep.equal(expectedData);
-    });
-
-    it('should set address and addressValidity correctly event without http', () => {
-      const validURL = '127.0.0.1:8080';
-      const data = wrapper.instance().validateUrl(validURL);
-      const expectedData = {
-        address: validURL,
-        addressValidity: '',
-      };
-      expect(data).to.deep.equal(expectedData);
-    });
-
-    it('should set address and addressValidity="URL is invalid" for a valid address', () => {
-      const validURL = 'http:localhost:8080';
-      const data = wrapper.instance().validateUrl(validURL);
-      const expectedData = {
-        address: validURL,
-        addressValidity: 'URL is invalid',
-      };
-      expect(data).to.deep.equal(expectedData);
-    });
-  });
-
   describe('changeHandler', () => {
     it('call setState with matching data', () => {
       wrapper = shallow(<Login {...props}/>, options);
